Refactor: 수정 영회님 코멘트 부분
- Board 엔티티 필드에 String 타입 deletedYn 필드 추가
- 게시 글 삭제 검증을 기존 querydsl이 아닌 Data Jpa
  existsByIdAndPasswordAndDeletedYn 쿼리 메서드를 활용해서 구현.
- 기능 수정에 따른 전반적인 코드 수정.
- 테스트 코드 수정.

Resolve: giantstep-study-board #7

+++ src/main/java/com/giantstep/board/domain/board/dto/BoardUpdateCheckRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 @Builder
 public class BoardUpdateCheckRequest {
-    private Integer boardId;
+    private Long boardId;
     private String boardPassword;
 }
